# monolith/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Documentation for GoWallet",
        "title": "GoWallet Monolith API",
        "contact": {
            "name": "API Support",
            "email": "bashocode@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/wallets/me": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get current authenticated user's wallet info (balance, currency, etc.)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Wallets"
                ],
                "summary": "Get My Wallet",
                "responses": {
                    "200": {
                        "description": "Returns success: true and data: model.Wallet",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_bashocode_gowallet_monolith_internal_errors.AppError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_bashocode_gowallet_monolith_internal_errors.AppError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_bashocode_gowallet_monolith_internal_errors.AppError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer \u003cyour_token\u003e\" to authenticate.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}